EmotionTag_forVideos/requirements.txt
opencv-python
python-docx
moviepy
pydub
SpeechRecognition
whisper
nltk
wordcloud
matplotlib